refactor: Removes dead code of Word Cloud plugin
