+++ superset-frontend/plugins/plugin-chart-word-cloud/src/index.ts
@@ -18,7 +18,5 @@
  */
 
 export { default as WordCloudChartPlugin } from './plugin';
-export { default as WordCloudTransformProps } from './plugin/transformProps';
-export { default as LegacyWordCloudChartPlugin } from './legacyPlugin';
 export * from './types';
 export { default as configureEncodable } from './configureEncodable';
