--- superset-frontend/plugins/plugin-chart-word-cloud/src/legacyPlugin/index.ts
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { t, ChartMetadata, ChartPlugin } from '@superset-ui/core';
-import transformProps from './transformProps';
-import buildQuery from '../plugin/buildQuery';
-import thumbnail from '../images/thumbnail.png';
-import { LegacyWordCloudFormData } from './types';
-
-const metadata = new ChartMetadata({
-  credits: ['https://github.com/jasondavies/d3-cloud'],
-  description: '',
-  name: t('Word Cloud'),
-  thumbnail,
-  useLegacyApi: true,
-});
-
-export default class LegacyWordCloudChartPlugin extends ChartPlugin<LegacyWordCloudFormData> {
-  constructor() {
-    super({
-      buildQuery,
-      loadChart: () => import('../chart/WordCloud'),
-      metadata,
-      transformProps,
-    });
-  }
-}
